Ports RMC mortar files and gets them into a half functional state
At the moment the game can build. The mortar and its ammo exist ingame and the mortar can be deployed. Trying to open the UI will call the open function but the UI does not actually open. Explosions are also entirely commented out and the code to support it have not been ported. Locale files are also not ported. All mortar related sprites have been ported.