Cleans the UI and unnecessary RMC dependencies
The blue horizontal line was removed as it would require editing base ss14 UI functions which I dont want to fuck with. The camera button on the mortar has been removed and the max dial section was moved above the dial buttons. RMC pop out window related code was removed.